8-16-18-pull: getting updated code from upstream master

+++ src/common/backend/BackendSystem.class.php
@@ -353,43 +353,48 @@ class BackendSystem extends Backend {
         $unix_group_name = $project->getUnixNameMixedCase();
         $ftp_anon_dir    = ForgeConfig::get('ftp_anon_dir_prefix')."/".$unix_group_name;
 
-        if (!is_dir($ftp_anon_dir)) {
-            // Now lets create the group's ftp homedir for anonymous ftp space
-            // This one must be owned by the project gid so that all project
-            // admins can work on it (upload, delete, etc...)
-            if (mkdir($ftp_anon_dir, 02775)) {
-                $this->chown($ftp_anon_dir, "dummy");
-                $this->chgrp($ftp_anon_dir, $unix_group_name);
-                chmod($ftp_anon_dir, 02775);
-            } else {
-                $this->log("Can't create project public ftp dir: $ftp_anon_dir", Backend::LOG_ERROR);
-                return false;
+        if (is_dir(ForgeConfig::get('ftp_anon_dir_prefix'))) {
+            if (!is_dir($ftp_anon_dir)) {
+                // Now lets create the group's ftp homedir for anonymous ftp space
+                // This one must be owned by the project gid so that all project
+                // admins can work on it (upload, delete, etc...)
+                if (mkdir($ftp_anon_dir, 02775)) {
+                    $this->chown($ftp_anon_dir, "dummy");
+                    $this->chgrp($ftp_anon_dir, $unix_group_name);
+                    chmod($ftp_anon_dir, 02775);
+                } else {
+                    $this->log("Can't create project public ftp dir: $ftp_anon_dir", Backend::LOG_ERROR);
+                    return false;
+                }
             }
+        } else {
+            $this->log("Skip create project public ftp dir: $ftp_anon_dir", Backend::LOG_INFO);
         }
-
         return true;
     }
 
     private function createProjectFRSDirectory(Project $project)
     {
         $unix_group_name = $project->getUnixNameMixedCase();
+        $ftp_frs_dir     = ForgeConfig::get('ftp_frs_dir_prefix')."/".$unix_group_name;
 
-        $ftp_frs_dir     = $GLOBALS['ftp_frs_dir_prefix']."/".$unix_group_name;
-
-        if (!is_dir($ftp_frs_dir)) {
-            // Now lets create the group's ftp homedir for anonymous ftp space
-            // This one must be owned by the project gid so that all project
-            // admins can work on it (upload, delete, etc...)
-            if (mkdir($ftp_frs_dir,0771)) {
-                chmod($ftp_frs_dir, 0771);
-                $this->chown($ftp_frs_dir, "dummy");
-                $this->chgrp($ftp_frs_dir, $this->getUnixGroupNameForProject($project));
-            } else {
-                $this->log("Can't create project file release dir: $ftp_frs_dir", Backend::LOG_ERROR);
-                return false;
+        if (is_dir(ForgeConfig::get('ftp_frs_dir_prefix'))) {
+            if (!is_dir($ftp_frs_dir)) {
+                // Now lets create the group's ftp homedir for anonymous ftp space
+                // This one must be owned by the project gid so that all project
+                // admins can work on it (upload, delete, etc...)
+                if (mkdir($ftp_frs_dir, 0771)) {
+                    chmod($ftp_frs_dir, 0771);
+                    $this->chown($ftp_frs_dir, "dummy");
+                    $this->chgrp($ftp_frs_dir, $this->getUnixGroupNameForProject($project));
+                } else {
+                    $this->log("Can't create project file release dir: $ftp_frs_dir", Backend::LOG_ERROR);
+                    return false;
+                }
             }
+        } else {
+            $this->log("Skip create project file release dir: $ftp_frs_dir", Backend::LOG_INFO);
         }
-
         return true;
     }
 
@@ -491,33 +496,27 @@ class BackendSystem extends Backend {
      * 
      * @return bool the status
      */
-    public function cleanupFRS() {
+    public function cleanupFRS()
+    {
+        $status = true;
         // Purge all deleted files older than 3 days old
-        if (!isset($GLOBALS['sys_file_deletion_delay'])) {
-            $delay = 3;
-        } else {
-            $delay = intval($GLOBALS['sys_file_deletion_delay']);
-        }
-        $time = $_SERVER['REQUEST_TIME'] - (3600*24*$delay);
-        $frs = $this->getFRSFileFactory();
-        $status =  $frs->moveFiles($time, $this);
-        // {{{ /!\ WARNING HACK /!\
-        // We keep the good old purge mecanism for at least one release to clean
-        // the previously deleted files
-        // Delete all files under DELETE that are older than 10 days
-        //$delete_dir = $GLOBALS['ftp_frs_dir_prefix']."/DELETED";
-        //system("find $delete_dir -type f -mtime +10 -exec rm {} \\;");
-        //system("find $delete_dir -mindepth 1 -type d -empty -exec rm -R {} \\;");
-        // }}} /!\ WARNING HACK /!\
+        $delay = (int) ForgeConfig::get('sys_file_deletion_delay', 3);
+        $time  = $_SERVER['REQUEST_TIME'] - (3600*24*$delay);
 
-        //Manage the purge of wiki attachments
-        $wiki   = $this->getWikiAttachment();
-        $status = $status && $wiki->purgeAttachments($time);
+        if (is_dir(ForgeConfig::get('ftp_frs_dir_prefix'))) {
+            $frs = $this->getFRSFileFactory();
+            $status =  $frs->moveFiles($time, $this);
+        }
+
+        if (is_dir(ForgeConfig::get('sys_wiki_attachment_data_dir'))) {
+            $wiki   = $this->getWikiAttachment();
+            $status = $status && $wiki->purgeAttachments($time);
+        }
 
         $em = EventManager::instance();
         $em->processEvent('backend_system_purge_files', array('time' => $time));
 
-        return ($status);
+        return $status;
     }
 
     /**
